backend/requirements.txt
# Django & API
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-environ==0.11.2
django-filter==23.3

# Database
psycopg2-binary==2.9.7
django-extensions==3.2.3

# Authentication & Security
djangorestframework-simplejwt==5.3.0
django-allauth==0.57.0

# File handling & OCR
Pillow==10.0.1
pytesseract==0.3.10
pdf2image==1.16.3
python-magic==0.4.27

# AI & ML
openai==1.3.5
langchain==0.0.340
langchain-community==0.0.1
tiktoken==0.5.1

# Async tasks
celery==5.3.4
redis==5.0.1
django-celery-beat==2.5.0
django-celery-results==2.5.1

# Search
django-haystack==3.2.1

# API Documentation
drf-spectacular==0.26.5

# Development
django-debug-toolbar==4.2.0
black==23.10.1
flake8==6.1.0
pytest==7.4.3
pytest-django==4.6.0
factory-boy==3.3.0

# Production
gunicorn==21.2.0
whitenoise==6.6.0
